Sliding Door / Fixed Spikes
Spikes are now triggers to avoid the issue of needing to be able to jump on them while also disabling it's collision with the player. Since the groundcheck relies on checking for the ground layer mask and the iFrames rely on checking for the enemy layer, I just made the spikes a trigger so the player could jump using the ground underneath the spikes but still cause the iFrames to trigger by using the enemy layer.