Flipping Player
The player Sprite flips depending on it's direction. The circle sprite to indicate the flip is temporary. I also fixed the enemy bugs and made it fall on a layer over top of the ground to emulate how Mario enemies die. I polished the knockback effect to avoid bugs like being launched high while already jumping adding an unintended bounce.